Add GetBattery command, query emergency backup battery state on boot and every 55 minutes (Sec+2.0 only) https://github.com/ratgdo/homekit-ratgdo/issues/302
I observed that my Sec+2.0 wall panel was sending a unknown command together with a GetOpenings to the
garage door every 22hrs 55mins.  The GDO was responding with both the cycle count and the state of the
emergency backup battery.  The unknown command was 0x09C which I have named GetBattery (observe that
battery state is reported with command 0x09D). I have added support for GetBattery and I call this during
startup (after first status is received) and every 55 minutes after that (as 22:55hrs seems like such
a long time).